Fix JWT createToken and validateToken functions
- Bcrypt hashed password is now stored in jwt payload.
- Stored password in jwt payload is compared against the password in the database to ensure that the access token does not remain valid if the password has since been changed.
- Console logs have been added to log:
i) the information of the requestor, such as: the IP of the client making the request; the endpoint that is being requested; and the access token and body/query parameters that were provided.
ii) the status of the token validation process, i.e. could the access token be verified, was the user authenticated, etc.